Implementing HashMap-based probability
This class now considers the number of times each phrase has been picked from a Serialized hashmap.

This class now has the Map<Integer, Integer> map as a frequency table that uses the phrase id and the number of appearances for the phrase.  The method setSayingMapDependent(Random rand) calculates the probability for each phrase.  The probability of appearance is inversely proportional to how many times its appeared.  updateMap() updates the map while initMap() instantiates the map for current/new save files.